# common/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseForeshadowing": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Foreshadowing"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "ApiResponseForeshadowingPage": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ForeshadowingPage"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Foreshadowing": {
        "description": "복선 카드 한 장. 회수 칸 셋(status · resolvedChapter · resolution)은 독자가 읽은 회차 N 기준으로 가린 값이다",
        "properties": {
          "chapter": {
            "description": "복선을 심은 회차",
            "example": 12,
            "format": "int32",
            "type": "integer"
          },
          "excerpt": {
            "description": "장부의 원문(영어)",
            "type": "string"
          },
          "fact": {
            "description": "카드 본문",
            "type": "string"
          },
          "id": {
            "description": "카드 번호. 담기·상세·판정 요청의 열쇠",
            "example": "T12",
            "type": "string"
          },
          "kind": {
            "description": "유형의 한국어 이름",
            "example": "약속",
            "type": "string"
          },
          "people": {
            "description": "인물 이름. 없으면 빈 배열",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "resolution": {
            "description": "회수 기록의 글. 미회수(또는 N화 뒤 회수)면 null",
            "type": "string"
          },
          "resolvedChapter": {
            "description": "회수된 회차. 미회수(또는 N화 뒤 회수)면 null",
            "example": 214,
            "format": "int32",
            "type": "integer"
          },
          "scene": {
            "description": "연결된 장면의 번호. 없으면 null",
            "example": "V12",
            "type": "string"
          },
          "sceneExcerpt": {
            "description": "장면의 글. 없으면 빈 글",
            "type": "string"
          },
          "status": {
            "description": "open 또는 resolved. N화 뒤에 회수된 복선은 open 으로 온다",
            "enum": [
              "open",
              "resolved"
            ],
            "type": "string"
          },
          "title": {
            "description": "카드 제목",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ForeshadowingPage": {
        "description": "카드 목록 한 쪽. chapter · page · size 는 요청한 값 그대로다 — 늦게 온 응답을 화면이 버리는 기준",
        "properties": {
          "chapter": {
            "description": "요청한 회차 N",
            "example": 200,
            "format": "int32",
            "type": "integer"
          },
          "chapterTotal": {
            "description": "N화 카드의 전체 수(검색어와 유형을 걸기 전)",
            "example": 774,
            "format": "int64",
            "type": "integer"
          },
          "hasNext": {
            "description": "다음 쪽이 있나",
            "type": "boolean"
          },
          "items": {
            "description": "카드. 순서는 표의 id 순 = T 번호 순",
            "items": {
              "$ref": "#/components/schemas/Foreshadowing"
            },
            "type": "array"
          },
          "page": {
            "description": "요청한 쪽. 0부터",
            "example": 0,
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "description": "요청한 크기",
            "example": 50,
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "description": "찾은 카드의 수(검색어와 유형을 건 뒤)",
            "example": 12,
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "description": "로그인하면 서버가 내려주는 HttpOnly 쿠키 `access_token`. 브라우저가 알아서 실어 보내므로 화면에서 따로 붙일 것이 없다. (앱처럼 쿠키를 쓰기 어려운 곳은 `Authorization: Bearer <토큰>` 도 받는다.)",
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Lore 창작 플랫폼 API 문서 (zzal / webtoon / piece-maker)",
    "title": "Lore API",
    "version": "v0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/piece-maker/v1/public/cards": {
      "get": {
        "description": "독자가 읽은 회차 N 이하에 심은 카드를 표의 id 순(= T 번호 순)으로 나눠 준다.\nN화 뒤에 회수된 복선은 status=open, resolvedChapter=null, resolution=null 로 가려서 준다.\n\n검색 규칙은 화면이 브라우저에서 찾던 것과 같다.\n· search 가 T12 꼴이면 그 번호의 카드 한 장만(대소문자 무관)\n· 그 밖에는 빈칸으로 나눈 단어가 모두 카드의 글(id · 제목 · 본문 · 인물 · 유형 · \"12화\")에 들어 있어야 한다.\n  소문자로 바꾸고 빈칸을 없애고 견준다. % 와 _ 는 글자 그대로 찾는다\n· kind 는 유형의 한국어 이름이 똑같은 카드만. 검색어와 함께 건다\n\nchapter 가 없거나 숫자가 아니거나 1~maxChapter 를 벗어나면 400(PIECE_MAKER_INVALID_CHAPTER).\nsearch 가 200자 또는 단어 10개를 넘거나 page · size 가 틀리면 400(INVALID_INPUT).",
        "operationId": "list_3",
        "parameters": [
          {
            "description": "독자가 읽은 회차 N. 1부터 maxChapter 까지",
            "example": 200,
            "in": "query",
            "name": "chapter",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "검색어. 200자 · 단어 10개까지. 비면 N화 카드 전부",
            "example": "밀짚모자",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "유형의 한국어 이름. 비면 거르지 않는다",
            "example": "약속",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "쪽. 0부터. 기본 0",
            "example": 0,
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "한 쪽의 크기. 기본 50, 최대 100",
            "example": 50,
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseForeshadowingPage"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "카드 목록과 검색",
        "tags": [
          "복선 카드(공개)"
        ]
      }
    },
...
    "/api/piece-maker/v1/public/cards/{id}": {
      "get": {
        "description": "카드 한 장을 번호로 연다. 게시글에 적힌 번호로 카드를 열 때 부른다.\n목록이 준 카드에는 상세의 칸이 다 있어서, 목록에서 여는 상세는 이 API 를 부르지 않는다.\n\n회수 칸 셋은 목록과 같은 규칙으로 N 기준으로 가린다.\n모르는 번호와 N화 뒤에 심은 카드는 같은 404(PIECE_MAKER_CARD_NOT_FOUND)다 — 구분해 주면\n번호를 바꿔 가며 뒤 회차의 카드가 있는지 알아낼 수 있다.",
        "operationId": "detail_1",
        "parameters": [
          {
            "description": "카드 번호",
            "example": "T12",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "독자가 읽은 회차 N. 1부터 maxChapter 까지",
            "example": 200,
            "in": "query",
            "name": "chapter",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseForeshadowing"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "카드 상세",
        "tags": [
          "복선 카드(공개)"
        ]
      }
    },
...
  },
...
}